refactor: Make `plugins` a module instead of a package
There's only one module in the package, we can always switch back to a package if needed later
