+++ narwhals/translate.py
@@ -35,7 +35,7 @@ from narwhals.dependencies import (
     is_pyarrow_scalar,
     is_pyarrow_table,
 )
-from narwhals.plugins._utils import _is_native_plugin, discover_entrypoints
+from narwhals.plugins import _is_native_plugin, discover_entrypoints
 
 if TYPE_CHECKING:
     from narwhals.dataframe import DataFrame, LazyFrame
